Factor compiler into its various components
